test: correct test for custom typescript config
As written, this test will pass even without the user settings to treat `.js` as Typescript, because TS is already set to use the .js extension in the LanguageConfig. Changing that to `.ts` causes this test to fail w/o the user config, and still pass with the user config.

While here, I also added some assertions to confirm that the default settings for as expected.