fix: match user custom file types by length
This doesn't allow custom settings to *always* win, but only when they are longer, or in a tie.

The recent change to use globs for user custom settings makes the UX of this a little fuzzy to me, though, because it muddies the idea of specificity: eg perhaps a suffix of `cc` should match as "more specific" than `c*`. In reality, though, I doubt this will really matter.

I haven't tested this commit yet, as I'm about to refactor the whole thing again anyway.